Break out tests across two files
